add user validity check
