+++ right-for-life_back/app.js
@@ -5,7 +5,7 @@ const homeRoutes = require('./routes/homeRoutes.js');
 const animalRoutes = require('./routes/animalsRoutes.js');
 const newsRoutes = require('./routes/newsRoutes.js');
 const happyStoriesRoutes = require('./routes/happyStoriesRoutes.js');
-const suppliesRoutes = require('./routes/supliesRoutes.js');
+const suppliesRoutes = require('./routes/suppliesRoutes.js');
 const aboutRoutes = require('./routes/aboutRoutes.js');
 const donateRoutes = require('./routes/donatesRoutes.js');
 const emergencyRoutes = require('./routes/emergenciesRoutes.js');
