+++ right-for-life_back/models/UserSchema.js
@@ -4,7 +4,7 @@ const UserSchema = new mongoose.Schema(
   {
     _id: mongoose.Schema.Types.ObjectId,
     email: { type: String, required: true, match: /[a-zA-Z0-9_\-.]+@[a-zA-Z]{1}[a-zA-Z.]+/g },
-    password: { type: String, required: true, match: /.{8,}/g },
+    password: { type: String, required: true, match: /[\w\W]{8,}/g },
     username: { type: String, required: true },
   });
