+++ right-for-life_back/routes/donatesRoutes.js
@@ -1,6 +1,7 @@
 const express = require('express');
 const router = express.Router();
 const mongoose = require('mongoose');
+const verifyUser = require('../utils/verifyUser.js');
 
 const DonateSchema = require('../models/DonateSchema.js');
 const DonateModel = mongoose.connection.model('Donate', DonateSchema);
@@ -15,6 +16,8 @@ router.get('/', (req, res, next) => {
 
 router.put('/', (req, res, next) => {
   const { title, manager, summary, paymentMethodsInfo, moneyTransferInfo } = req.body;
+
+  if (!verifyUser(JSON.parse(req.get('Authorization')))) res.status(401).end();
 
   DonateModel.findOne()
     .exec()
