+++ right-for-life_back/utils/verifyUser.js
@@ -1,0 +1,20 @@
+const jwt = require('jsonwebtoken');
+const secret = require('./configs.js').jwt_secret;
+
+function verify(authToken) {
+  const tokenExpiration = jwt.decode(authToken).exp * 1000;
+  const currentDate = Date.now();
+
+  if (currentDate > tokenExpiration) {
+    return false;
+  }
+
+  try {
+    jwt.verify(authToken, secret);
+    return true;
+  } catch (e) {
+    return false;
+  }
+}
+
+module.exports = verify;
